// src/app/models/user-bd-full.ts
export interface UserDbFull{
    APRESENTACAO: string;
    ATRIBUICOES: string;
    BAIRRO: string;
    CARGO: string;
    CEP: string;
    CIDADE: string;
    CONTATO: string;
    CPFCNPJ:string;
    CURSO:string;
    DEFICIENTE:string;
    DTALTERACAO: string;
    DTCONCLUSAO: string;
    DTINICIO:string;
    DTNASCIMENTO: string;
    DTSAIDA:string;
    EMAIL: string;
    EMPRESA: string;
    ESTADO: string;
    ESTADOCIVIL: string;
    FLAGATIVO: string;
    FLAGCONCLUIDO: string;
    IDEMDERECO: string;
    IDEXPERIENCIA: string;
    IDFORMACAO: string;
    TIPOFORMACAO: string;
    IDUSER: string;
    INSTITUICAO: string;
    LINKEDIN: string;
    LOGRADOURO: string;
    NOME: string;
    NUMERO: string;
    RECADO: string;
    SEXO: string;
    UFNASCIMENTO: string;
    PRETENSAOSALARIAL:number
}